Fix CI warnings for Windows Clang-CL and other platforms
- Add SDL2/SDL3 compatibility for SDL_GL_CONTEXT_FORWARD_COMPATIBLE
- Fix 32-bit shift warning in trilinear.hh with static_cast<size_t>
- Add Clang-specific warning suppressions for strict Clang-CL warnings
- Add _CRT_SECURE_NO_WARNINGS for Windows to fix getenv deprecation
- Move NOMINMAX to WIN32 block to apply to both MSVC and Clang-CL

+++ include/scaler/cpu/trilinear.hh
@@ -102,9 +102,9 @@ namespace scaler {
             }
 
             // Each level halves the dimensions
-            const size_t scale_divisor = 1 << level; // 2^level
-            const size_t mip_width = std::max(size_t(1), src.width() / scale_divisor);
-            const size_t mip_height = std::max(size_t(1), src.height() / scale_divisor);
+            const size_t scale_divisor = static_cast<size_t>(1) << level; // 2^level
+            const size_t mip_width = std::max(static_cast<size_t>(1), src.width() / scale_divisor);
+            const size_t mip_height = std::max(static_cast<size_t>(1), src.height() / scale_divisor);
 
             OutputImage result(mip_width, mip_height, src);
